api/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asttokens==2.4.1
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
diskcache==5.6.3
distro==1.9.0
docker==7.0.0
executing==2.0.1
fastapi==0.109.2
FLAML==2.1.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.7
ipython==8.19.0
jedi==0.19.1
Jinja2==3.1.3
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
numpy==1.26.3
openai==1.6.1
packaging==23.2
parso==0.8.3
pexpect==4.9.0
pillow==10.3.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pyautogen==0.2.31
pydantic==2.8.2
Pygments==2.17.2
PyJWT==2.8.0
python-dotenv==1.0.0
regex==2023.12.25
requests==2.32.2
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
starlette==0.36.3
termcolor==2.4.0
tiktoken==0.5.2
tqdm==4.66.3
traitlets==5.14.1
urllib3==2.2.2
uvicorn==0.30.1
wcwidth==0.2.13
